REFACTOR: Moving all manipulator-related code in SofaViewer
