+++ applications-dev/plugins/SofaQtQuickGUI/data/qml/SofaWidgets/EditModeManipulators.qml
@@ -12,6 +12,7 @@ import SofaManipulators 1.0
 
 Column {
     id: manipulatorControls
+    property var sofaViewer
     anchors.left: parent.left
     anchors.top: parent.top
     anchors.leftMargin: 20
@@ -19,13 +20,13 @@ Column {
 
 
     function setManipulator(manipulatorName) {
-        var m = SofaApplication.getManipulator(manipulatorName)
-        if (!SofaApplication.getManipulators()) return m
-        for (var i = 0 ; i < SofaApplication.getManipulators().length ; ++i) {
-            if (!SofaApplication.getManipulators()[i].persistent &&
-                    (m && m.name !== SofaApplication.getManipulators()[i].name)) {
-                SofaApplication.getManipulators()[i].enabled = false
-                print("disabling " + SofaApplication.getManipulators()[i].name)
+        var m = sofaViewer.getManipulator(manipulatorName)
+        if (!sofaViewer.manipulators) return m
+        for (var i = 0 ; i < sofaViewer.manipulators.length ; ++i) {
+            if (!sofaViewer.manipulators[i].persistent &&
+                    (m && m.name !== sofaViewer.manipulators[i].name)) {
+                sofaViewer.manipulators[i].enabled = false
+                print("disabling " + sofaViewer.manipulators[i].name)
             }
         }
         m.enabled = true
@@ -35,7 +36,7 @@ Column {
     ManipulatorMenu {
         id: translateMenu
         property string manipulatorName: "Translate_Manipulator"
-        manipulator: SofaApplication.getManipulator(manipulatorName)
+        manipulator: sofaViewer.getManipulator(manipulatorName)
 
         onOptionChanged: {
             manipulator = setManipulator(manipulatorName)
@@ -61,17 +62,20 @@ Column {
         id: rotateMenu
 
         property string manipulatorName: "Rotate_Manipulator"
-        manipulator: SofaApplication.getManipulator(manipulatorName)
+        manipulator: sofaViewer.getManipulator(manipulatorName)
 
         image: "qrc:/icon/ICON_ROTATION_MODIFIER.png"
         model: ListModel {
             ListElement {
                 title: "Local"
                 option: true
+                iconSource: "qrc:/icon/ICON_ROTATION_MODIFIER.png"
             }
             ListElement {
                 title: "Global"
                 option: false
+                iconSource: "qrc:/icon/ICON_ROTATION_WORLD_MODIFIER.png"
+
             }
         }
